+++ src/gwt/src/org/rstudio/core/client/theme/res/ThemeResources.java
@@ -262,4 +262,10 @@ public interface ThemeResources extends ClientBundle
    
    @Source("browsePackage_2x.png")
    ImageResource browsePackage2x();
+
+   @Source("suspended.png")
+   ImageResource suspended();
+
+   @Source("suspendBlocked.png")
+   ImageResource suspendBlocked();
 }