+++ src/gwt/src/org/rstudio/studio/client/projects/ui/newproject/NewQuartoBookProjectPage.java
@@ -17,13 +17,13 @@ package org.rstudio.studio.client.projects.ui.newproject;
 import com.google.gwt.core.client.GWT;
 import org.rstudio.core.client.resources.ImageResource2x;
 import org.rstudio.studio.client.projects.StudioClientProjectConstants;
-import org.rstudio.studio.client.quarto.model.QuartoConstants;
+import org.rstudio.studio.client.quarto.model.QuartoCommandConstants;
 
 public class NewQuartoBookProjectPage extends NewQuartoProjectPage
 {
    public NewQuartoBookProjectPage()
    {
-      super(QuartoConstants.PROJECT_BOOK,
+      super(QuartoCommandConstants.PROJECT_BOOK,
            constants_.quartoBookTitle(),
            constants_.quartoBookSubTitle(),
            constants_.quartoBookPageCaption(),