+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/CompletionContext.java
@@ -20,4 +20,8 @@ public interface CompletionContext
    String getId();
    String getPath();
    String getExtendedFileType();
+   String[] getQuartoFormats();
+   String[] getQuartoProjectFormats();
+   String getQuartoEngine();
+   
 }